src/test/java/com/rares/syneco/converter/SynConverterTest.java
package com.rares.syneco.converter;

import static org.junit.Assert.assertThat;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.reactive.server.WebTestClient;

import static org.assertj.core.api.Assertions.*;


@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
@RunWith(SpringRunner.class)
public class SynConverterTest {

	@Autowired
	private SynConverter converter;

	@Test
	public void exampleTest() {
		assertThat(converter.respondToInput("sf")).isEqualTo("salutsf");
	}
}
